+++ src/esrs/aggregate_manager.rs
