+++ src/esrs/mod.rs
@@ -1,10 +1,10 @@
 pub mod aggregate;
+pub mod aggregate_manager;
+pub mod aggregate_state;
 pub mod event_bus;
 pub mod event_handler;
-pub mod manager;
+pub mod event_store;
 pub mod rebuilder;
-pub mod state;
-pub mod store;
 
 #[cfg(feature = "postgres")]
 pub mod postgres;
